+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-08-29 23:08:46
-// Commit: feat(ui): add async ui stream provider
+// Updated at: 2026-08-31 20:26:09
+// Commit: feat(ui): integrate custom telemetry logger in ui
 
 class uiComponentState {
-  final String commitRef = '4057';
-  final String lastMessage = 'feat(ui): add async ui stream provider';
-  final DateTime timestamp = DateTime.parse('2026-08-29T23:08:46+05:00');
+  final String commitRef = '4095';
+  final String lastMessage = 'feat(ui): integrate custom telemetry logger in ui';
+  final DateTime timestamp = DateTime.parse('2026-08-31T20:26:09+05:00');
 }